save time by reading existing stations.csv

+++ 2021-05-27-weather-forecast/weather-forecast-table.R
@@ -22,11 +22,16 @@ doc <- XML::xmlParse(kml)
 datetime <- get_datetime(doc)
 meta     <- get_meta_info(doc)
 
-stations <- read_fwf('https://www.dwd.de/EN/ourservices/met_application_mosmix/mosmix_stations.cfg;jsessionid=6A47F126BB1F3B5383F07B7CC03595E9.live31092?view=nasPublication&nn=495490',
-                      col_positions = fwf_widths(c(5,6,6,5,21,7,8,6,7,5),
-                                                 col_names = c('clu','CofX','id','ICAO', 'name','lat','lon','elev','Hmod-H', 'type')), skip = 4)
 
-good.stations <- stations %>% filter(!is.na(as.numeric(id)) & type=='LAND')
+if (file.exists('out/stations.csv')) {
+  good.stations <- read_csv('out/stations.csv')
+} else {
+  stations <- read_fwf('https://www.dwd.de/EN/ourservices/met_application_mosmix/mosmix_stations.cfg;jsessionid=6A47F126BB1F3B5383F07B7CC03595E9.live31092?view=nasPublication&nn=495490',
+                       col_positions = fwf_widths(c(5,6,6,5,21,7,8,6,7,5),
+                                                  col_names = c('clu','CofX','id','ICAO', 'name','lat','lon','elev','Hmod-H', 'type')), skip = 4)
+  good.stations <- stations %>% filter(!is.na(as.numeric(id)) & type=='LAND')  
+}
+
 
 final_stations <- c()
 
@@ -57,7 +62,9 @@ for (station_id in good.stations$id) {
   }
 }
 
-good.stations %>% filter(id %in% final_stations) %>% write_csv('out/stations.csv')
+if (!file.exists('out/stations.csv')) {
+  good.stations %>% filter(id %in% final_stations) %>% write_csv('out/stations.csv')
+}
 
 f <- file('out/last-update.txt')
 writeLines(as.character(Sys.time()), f)
